trying easier structure for cloning/submodules
